+++ CHANGES.rst
@@ -8,3 +8,6 @@ v0.1.0
 * Update README
 * Add initial migration
 -----------
+v0.2.0
+* Implement model and serializer logic
+* Implement views, helpers and emails